removing unncessary files
