--- Midas/data_preprocessing/Dockerfile
@@ -1,10 +1,0 @@
-FROM amancevice/pandas:latest
-
-ADD requirements.txt ./requirements.txt
-
-RUN pip install --upgrade pip \
-    && pip install --no-cache-dir -r ./requirements.txt
-
-COPY main.py ./main.py
-
-ADD . .
