--- Midas/data_preprocessing/Makefile
@@ -1,7 +1,0 @@
-all: pull run
-
-pull:
-	docker build -t data_preprocessing .
-
-run:
-	docker run --network=databases_default data_preprocessing
