--- Midas/data_preprocessing/README.md
@@ -1,5 +1,0 @@
-## Getting Started with Docker
-
-1. Sign up for an account at `https://hub.docker.com/`.
-2. Install the appropriate docker application for your desktop.
-3. In this directory, run `make`. This should pull the docker image, build it, and run the job associated with the docker container.
